fix(ui): LocalizedCurrencyFormatter accepts Decimal to match ProductCandidateMatch.estimatedPrice

+++ AssetCore/Sources/UIComponents/ProductConfirmationModal.swift
@@ -193,7 +193,7 @@ public struct ProductConfirmationModal: View {
 
 @MainActor
 private enum LocalizedCurrencyFormatter {
-    static func format(amount: Double, currencyCode: String) -> String {
-        RegionalFormatter.shared.formatCurrency(amount: Decimal(amount), currencyCode: currencyCode, locale: LanguageManager.shared.currentLocale)
+    static func format(amount: Decimal, currencyCode: String) -> String {
+        RegionalFormatter.shared.formatCurrency(amount: amount, currencyCode: currencyCode, locale: LanguageManager.shared.currentLocale)
     }
 }
